LayoutTests/inspector-protocol/debugger/setBreakpoint-dfg-callee-and-examine-dfg-local.html

<html>
<head>
<script src="../../http/tests/inspector-protocol/resources/protocol-test.js"></script>
<script src="resources/breakpoint.js"></script>

<script>
// Put this here instead of on <body onload> to prevent an extra Debugger.scriptParsed event.
window.onload = runTest;

function test()
{
    // This test setting a breakpoints in DFG compiled functions callee and then modify
    // and examine a global and local via the DFG frame.

    InspectorTest.sendCommand("Debugger.enable", {});

    var breakpointId = null;
    var scriptId = 0;
    var startLine = 0;

    InspectorTest.eventHandler["Debugger.scriptParsed"] = function(messageObject)
    {
        if (/resources\/breakpoint\.js$/.test(messageObject.params.url)) {
            InspectorTest.log("Found breakpoint.js");
            scriptId = messageObject.params.scriptId;
            startLine = messageObject.params.startLine;

            InspectorTest.sendCommand("Runtime.evaluate", {
                expression: "dfgWithoutInline3();"
            }, function(responseObject) {
                InspectorTest.log("dfg function warmed up\n");

                var location1 = {scriptId: scriptId, lineNumber: 2, columnNumber: 0};

                InspectorTest.sendCommand("Debugger.setBreakpoint", {location: location1}, function(responseObject) {
                    InspectorTest.checkForError(responseObject);
                    InspectorTest.log("Breakpoint set in breakpointBasic()");

                    breakpointId = responseObject.result.breakpointId;
                    InspectorTest.sendCommand("Runtime.evaluate", {
                        expression: "dfgWithoutInline3();"
                    }, function(responseObject) {
                        InspectorTest.log("Test complete");
                        InspectorTest.completeTest();
                    });
                });
            });
        }
    }

    InspectorTest.eventHandler["Debugger.paused"] = function(messageObject)
    {
        function dumpResponse(response)
        {
            try {
                if (response.result.wasThrown) {
                    InspectorTest.log("Exception thrown processing request");
                    return false;
                }
                InspectorTest.log("Response value is " + response.result.result.value);
                return true;
            } catch (e) {
                return false;
            }
        }

        function resumeFromBreakpoint()
        {
            InspectorTest.sendCommand("Debugger.resume", {}, function(responseObject) {
                InspectorTest.log("Resumed from breakpoint");
            });
        }

        InspectorTest.log("Hit Breakpoint!");
        var callFrames = messageObject.params.callFrames;
        if (callFrames.length < 3) {
            InspectorTest.log("FAIL: too few frames in stack trace");
            resumeFromBreakpoint();
            return;
        }
        var callFrameId = callFrames[2].callFrameId;
        InspectorTest.log("Evaluating in DFG frame at frame[2]: 'globalVal3 = 30;'");
        InspectorTest.sendCommand("Debugger.evaluateOnCallFrame", { callFrameId: callFrameId, expression: "globalVal3 = 30;" }, function(responseObject) {
            if (!dumpResponse(responseObject)) {
                resumeFromBreakpoint();
                return;
            }
            InspectorTest.log("Evaluating in DFG frame at frame[2]: 'localVal3 = 12;'");
            InspectorTest.sendCommand("Debugger.evaluateOnCallFrame", { callFrameId: callFrameId, expression: "localVal3 = 12;" }, function(responseObject) {
                if (!dumpResponse(responseObject)) {
                    resumeFromBreakpoint();
                    return;
                }
                InspectorTest.log("Evaluating in DFG frame at frame[2]: 'localVal3'");
                InspectorTest.sendCommand("Debugger.evaluateOnCallFrame", { callFrameId: callFrameId, expression: "localVal3" }, function(responseObject) {
                    if (!dumpResponse(responseObject)) {
                        resumeFromBreakpoint();
                        return;
                    }
                    InspectorTest.sendCommand("Debugger.resume", {}, function(responseObject) {
                        InspectorTest.log("Resumed from breakpoint");
                    });
                });
            });
        });
    }
}
</script>
</head>
<body>
<p>Debugger.evaluateOnCallFrame in a DFG compiled function from a breakpoint in a non-DFG callee.</p>
</body>
</html>
